backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pillow>=10.2.0
rembg>=2.0.50
python-multipart>=0.0.6
pydantic>=2.5.0
mediapipe>=0.10.9
numpy>=1.26.0
